Compile and Import Player sprites
There are 16 sprites assigned to the player. They should be stored in dist/assets/spritesheets and drawn in greyscale as a single column of 16 8x8 tiles: The sprites are processed by dividing the r component by 85 and using that result to select a color in [0-3] from the palette assigned to the player.